control-api/requirements.txt
anyio==3.5.0
asgiref==3.5.0
certifi==2021.10.8
chardet==4.0.0
click==8.1.2
fastapi==0.75.1
h11==0.13.0
hjson==3.0.2
hologram-python==0.9.1
idna==2.10
mock==3.0.5
psutil==5.8.0
pycryptodome==3.9.0
pydantic==1.9.0
pyroute2==0.5.19
pyserial==3.5
python-dotenv==0.20.0
python-pppd==1.0.3
python-sdk-auth==0.3.9
pyudev==0.22.0
pyusb==1.1.1
requests==2.25.1
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing-extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
